+++ integration-tester/tests/99-switch-namespace/switchNamespace.js
@@ -5,7 +5,7 @@ const infra = infrastructure.createTestInfrastructure();
 
 const numContainers = Math.round(infrastructure.nWorker / 2);
 for (let i = 0; i < numContainers; i += 1) {
-  (new kelda.Container('testContainer', 'alpine', {
+  (new kelda.Container('test-container', 'alpine', {
     command: ['tail', '-f', '/dev/null'],
   })).deploy(infra);
 }