+++ CHANGELOG.md
@@ -4,6 +4,10 @@ Kelda Change Log
 Up Next
 -------------
 
+- Container and Load Balancer hostnames must now comply with the Kubernetes
+DNS_LABEL spec. Specifically, hostnames must contain only lower case
+characters, numbers, or hyphens.
+
 Release 0.8.0
 -------------